refactor: adds lost changes due to merge conflicts
