+++ .gitignore
@@ -339,7 +339,7 @@ Sessionx.vim
 # Temporary
 .netrwhist
 # Auto-generated tag files
-tags
+# tags
 # Persistent undo
 [._]*.un~
