+++ src/encord_active/app/model_quality/components/false_negative_view.py
@@ -8,18 +8,19 @@ from streamlit.delta_generator import DeltaGenerator
 import encord_active.app.model_quality.components.utils as cutils
 from encord_active.app.common import state
 from encord_active.app.common.components import build_data_tags
-from encord_active.app.common.components.bulk_tagging_form import (
+from encord_active.app.common.components.tags.bulk_tagging_form import (
     BulkLevel,
     action_bulk_tags,
     bulk_tagging_form,
 )
-from encord_active.app.common.components.individual_tagging import multiselect_tag
+from encord_active.app.common.components.tags.individual_tagging import multiselect_tag
 from encord_active.app.common.utils import (
     build_pagination,
     get_df_subset,
     load_or_fill_image,
 )
 from encord_active.lib.common.colors import Color, hex_to_rgb
+from encord_active.lib.metrics.load_metrics import MetricScope
 
 
 def __show_image_and_fn(
@@ -59,7 +60,7 @@ def __build_card(
 ):
     with st_col:
         __show_image_and_fn(label, predictions, box_color=box_color)
-        multiselect_tag(label, "false_negatives", MetricType.MODEL_QUALITY)
+        multiselect_tag(label, "false_negatives", MetricScope.MODEL_QUALITY)
 
         cls = st.session_state.full_class_idx[str(label["class_id"])]["name"]
         label = label.copy()
@@ -77,7 +78,7 @@ def false_negative_view(false_negatives, model_predictions, color: Color):
     subset = get_df_subset(false_negatives, selected_metric)
     paginated_subset = build_pagination(subset, n_cols, n_rows, selected_metric)
 
-    form = bulk_tagging_form(MetricType.MODEL_QUALITY)
+    form = bulk_tagging_form(MetricScope.MODEL_QUALITY)
 
     if form and form.submitted:
         df = paginated_subset if form.level == BulkLevel.PAGE else subset
